Move WordPress bottom sheet dialog to login library
