+++ WordPress/src/main/res/values/dimens.xml
@@ -294,8 +294,6 @@
 
     <dimen name="button_container_shadow_height">5dp</dimen>
 
-    <dimen name="bottom_sheet_dialog_width">0dp</dimen>
-
     <!-- signup -->
     <dimen name="progress_spinner_small">24dp</dimen>
